+++ lib/host_event.dart
@@ -1,7 +1,6 @@
 part 'host_events/request_crypto_account_result.dart';
 part 'host_events/send_crypto_result.dart';
 
-/// Outgoing host → widget events for the Flutter SDK surface.
 sealed class HostEvent {
   const HostEvent();
